+++ src/rastervision/filesystem/__init__.py
@@ -1,0 +1,6 @@
+# flake8: noqa
+
+from rastervision.filesystem.filesystem import FileSystem
+from rastervision.filesystem.local_filesystem import LocalFileSystem
+from rastervision.filesystem.s3_filesystem import S3FileSystem
+from rastervision.filesystem.http_filesystem import HttpFileSystem
