+++ src/rastervision/filesystem/filesystem.py
